Update TransactionListSummary and Radial Chart code (#404)
* Update TransactionListSummary and Radial Chart code

* Update Popper element styles and functionality

* Update all delete actions to have loader

+++ packages/frontend/src/actions/category/handleCategoryDelete.ts
@@ -3,9 +3,17 @@
 import { redirect, RedirectType } from 'next/navigation';
 
 import { settingsPaths } from '$constants/settings-paths';
+import { DefaultFormActionHandler } from '$hooks/useFinancerFormState';
 import { CategoryService } from '$ssr/api/category.service';
 
-export const handleCategoryDelete = async (id: string) => {
+export const handleCategoryDelete: DefaultFormActionHandler<{
+  id: string;
+}> = async ({ id }) => {
+  if (!id) {
+    return { status: 'ERROR', errors: ['Failed to delete category: no id'] };
+  }
+
   await CategoryService.delete(id);
+
   redirect(settingsPaths.categories, RedirectType.push);
 };
